requirements.txt
django>=3.2
pydantic
django-ninja
jinja2
tomlkit
pyyaml
uvicorn[standard]
httpx
typer
